Fix warning for 32 vs 64 devices

+++ LaunchKit/Classes/UI/Components/LKViewController.m
@@ -334,7 +334,7 @@ NSString * NSStringFromViewControllerFlowResult(LKViewControllerFlowResult resul
         case LKViewControllerFlowResultFailed:
             return @"failed";
         default:
-            NSLog(@"Couldn't understand flow result %ld; returning 'unknown'", result);
+            NSLog(@"Couldn't understand flow result %ld; returning 'unknown'", (long)result);
             return @"unknown";
     }
 }
